PIXIE-42: Revamping the command line argument parsing.
The new parser is much more robust than the old one. It's
pretty similar to argparse from python. We should be able to
handle the basics of what we need with this code. It also
creates the help functions for us so we don't need to deal
with all of that.