P16-63: Adding PCF8574 support
This commit adds support for the system to read the slot number from the
module's PCF8574 chip. This removes the need for users to have the pxisys.ini
file in their system. It also adds support for the new Wiener PCIe crates.

S. V. Paulauskas <stan@xia.com> made the commit, but the work was done by Hui.