P16-82: Updating the code so that we only build a single library instead of two
We now just build PixieSDK.so instead of PixieSys.so and PixieApp.so.
We've combined all these files into a single directory called sdk. This gives us
a single place to look for our files and make changes. This has also simplified
our build scripts as we no longer need to maintain twh different scripts.