P16-79: Updating logging, removing files, cleaning up implementation
Updated logging from cout statements to easylogging.

Moved some of the generally useful functions to the system_test_utilities.hpp file.

Added the ability for users to write once time but read multiple times.

Added the ability for users to do a sequential DSP read instead of burst.

Add the ability for users to execute the test more than once.