P16-179: Building two libraries now
We're now building both the LegacyPixieSdkObjLib and PixieSdkObjLib.
PixieSDK.so contains both object libraries to ease transitions to the new system.
We keep LegacyPixieSDK.so. Once we rename the pixie16.cpp functions to match
the old API, we'll end up with conflicts.
This will make it easier to separate the two later.