P16-222: Fixing issue with constexpr using sqrt
sqrt isn't defined as a constexpr, it can't be used to initialize a constexpr variable.
Removed the expr part and it works fine.

I also removed the custom definition of pi and used the cmath version of pi / 2.

Windows won't recognize the variable name without defining _USE_MATH_DEFINES.
I added the define to the windows compat header.