P16-160: level_logging now returns properly for off levels
We expect that a level_logging would always return true for the off level.
This level acts like a sink for messages and thus will always be available.

If a logger is set to off, then a call to level_logging checking any flag other
than off returns false. Those levels are all deactivated in this scenario.