P16-18: Fixing test so that it passes the check we're executing.
TAU writes were failing because the parameter write executes a BL cut update.
This isn't implemented in the simulation and so it was throwing an error. The
actual write is performing as expected, which is what we want to test here.
We've added in a simple try/catch to catch the error, and then verify that the
write happened with the asserts. We can remove the try/catch after we simulate
control tasks.