P16-83: Temp files use tmpnam and adding catch for invalid bit values
The Set/Clr/Tst bit functions did not behave consistently across windows and linux.
To ensure that we get consistent behavior we implemented a catch to ensure that
users cannot request a bit greater than what we want.
For example, APP16_ClrBit is expecting the bit to be between 0 and 15.
If the value is greater than 15, then we return the input value, as nothing
should change.

We have also updated the test suite to use the built in function for generating
temporary file names. This ensures that pathing is correct for the platform.