P16-290: Massive organization of files to improve readability
We've gotten rid of the API directory in favor of keeping all our includes
together. The new include organization matches the installation paths. This improves
our customer's ability to locate files in the project. The pathing simplifies
the installation, as we can just install the include directory as is.

The new organization allows us to expand to additional pixie products in the
future. We can expand within the pixie/ directory as needed. For example, we can add a pixie/pixie-net directory to capture updates for a pixie-net.