P16-300: Added options to turn on/off various parts of the SDK.
Added some options as dependent options. For example, we can’t build the
Pixie16 support library (Pixie16api.so) without also building the SDK. If a
user sets the SDK to OFF then the dependent flags will be disabled
and turned off.

Updated the various builds to be dependent on the appropriate flags.