P16-346: Updating example_pixie16api.cpp to use JSON configs
I've updated the code so that it uses a JSON configuration file. The old file
didn't read in properly when there were spaces in the file name.

I've also revamped functions to accommodate the new data structures.

Took the opportunity to rename the folder to be consistent with the file naming.

Updated the README for the examples to now be consistent with functionality.