P16-196: Updating examples to handle stats data output.
I'm seeing some issues with the statistics data in the glue code. To help with
debugging and visualization, we now output stats data in binary and csv.

The example code now also outputs the statistics data for MCA runs.

I've updated data_visualizer.py to handle the stats CSV output. It creates
a bar chart with the ICR and OCR data.

data_visualizer.py now outputs a PNG image with the graph for users who don't
use some sort of IDE.