P16-287: Fixing bug introduced by alphabetizing enum entries.
We use the variable and parameter enums as lookup tables when coupled with the
descriptor maps. That means that the order in the enum must be the same as
the order in the descriptor maps.

I've updated the comments for these enums to tell us that not only does the
order matter, but if you're adding a new entry, then you must update teh
corresponding data structures.