P16-360: As close to gcc 4.9.2 support that we're going to get.
We're keeping the CMake revision at 3.10. We want to make sure that we're not
continuing to support older systems forever.

NOTE that the unit tests don't compile with 4.9.2. Doctest needs FULL C++ 11 support.

Cleaned up xia_configure_target.cmake so that we set the CXX standard on a
per-target basis. This ensures that we can change that later if we need.
Defaults to 14.

Fixed issue in FindBroadcomAPI.cmake and FindPixieSDK.cmake that prevented
CMake 3.0.2 from setting properties properly.

Needed to change just a bit of code in a few spots to accommodate the lack of
full C++14 support in 4.9.2. The changes don't impact performance in an appreciable way.