P16-377: Implements Pixie16LoadDSPParametersFromFile
I dried out the load settings calls by creating a function that we use in
Pixie16BootModule and Pixie16LoadDSPParametersFromFile.

I modified crate::import_config to take an additional flag to control
synchronizing the parameters to the hardware. The flag defaults to true.
We needed this flag to allow calls to import_config that just affected the
local memory.

Added the --additional-config flag to example_pixie16api.cpp that takes an
 additional configuration file. If this flag is present, then the code will
 execute a fast boot, and make a call to Pixie16LoadDSPParametersFromFile
 with the flag's argument.