P16-381: Fixing messaging and checks for chip boot status
The module::boot function was using the wrong variable for the SP FPGA
when outputting the status of the chips. I updated it to use the correct one.

I also updated a check in module::probe where the DSP check was dependent on
the SP FPGA. You can boot the DSP without booting the SP FPGA so that
dependency wasn't correct.