P16-214: Fixing BLCut finder implementation
We just needed to change baseline::compute_cut so that it used an indexed
for loop instead of a range based loop. The ranged loop used the channel
number to index the vectors. This caused issues since for BLCut finder these
vectors all have a size of 1. Changing this to be indexed fixes the issue.