P16-371: Removing CRC check on settings file
This implementation forces us to export functions from PixieSDK.a for Windows
to be happy. We want to avoid this if at all possible.

We'll also need to be able to change this file in the future. The file being
checked only works for the specific setup in place now. We may also need to
modify the filter settings or thresholds.

I updated the settings file basename to be the same as the test. That should
ensure that we don't run into the issue with generically named files.