P16-410: Updating warning and default config
I moved the channel extension code to be guarded by a check on channels.
We don't always need to execute that section of code.

Modified the extension warning to be debug statement.

Updated the default configuration object to only have a single array element.
The new implementation will fill out the rest of the vectors automatically.

Warning about too few elements now only prints if it's not the default configuration.