P16-418: Fixing API call for reading channel histograms.
There was a bug that allowed users to request more data than the histogram
contained. We now truncate the read if the user requests more data than the
channel fixture has stored as its max. This also fills the truncated values
with the max value of an unsigned int.

We've also added a catch for users that request less than the maximum.
In this case they may be lead to believe that they are missing data when they
just didn't read it all.