P16-427: Removing remaining crate.ready calls.
Many of the module level functions use the ModNum == crate.num_modules check
to see if we need to run the command against every module in the system
in those cases the module handles weren't getting called properly and could
cause issues. We've now cleaned up the usage of the module handle where possible.