P16-441: Moving Tau Finder to module level
This doesn't get the right data back yet, but I think everything's in place for the rest.

We don't want the API exposing control task runs. It's a matter of controlling
the access that users see. We needed to move the Tau finder functionality
to the module level and get it integrated with the fixtures.

The SDK will eventually need to take over this functionality.