P16-314: Implements an event class to store decoded list-mode events
We've stored the class within the xia::pixie::data::list_mode namespace.
The namespace keeps the readable nomenclature that we like, and ensures that
the event class is specific to this type of data.

This class allows us to store decoded list-mode data. We have provided it
with a json serializer and deseralizer.

This namespace creates a dedicated shared object. This is meant to be a
helper library that users can decide to use in their software. It's not
required for the SDK to work.