P16-463: Adding debugging and information for list-mode verification
We found that during testing we needed some additional logging output to
facilitate validation of the list-mode data fidelity and performance.

We also found that the legacy API needs to have some waits added in the
example. These waits seem to be necessary to make sure that the DSP has
started the data runs properly. Without them the data runs exit immediately.