P16-314: Removes any dependence of the firmware and freq from the data
We were presenting the users with the high/low components of the time.
This was a mistake since we only break those up due to the 32bit limitation
on the hardware. We now combine all time components that were previously
broken into a low and high.

We now store all times in units of seconds. This allows us to remove the last
dependence of the sampling and FPGA frequency from the data.
We cannot store the external time in seconds since we don't know the
clock frequency. Users will need to convert that themselves.