P16-472: Fixes data_visualizer.py for list-mode data
The way we calculate the padding doesn't catch all the edge cases.
We remove this for list-mode data as it doesn't make sense to calculate here.
We have to keep it for the adc and mca data, but those are generally going to
be better behaved than the list-mode data.